Serializer.convertShort: removed
replaced by 'C.pack . show'

The Haskell foreign import of cshortToString had wrong parameter type Int.
It should have been CShort.
But this silently wraps around any integer in a PDF document at 32767.

I tried to replace 'short' by 'long', but this does not solve the problem.
In order to prevent silent wrap-around
I would have to add a check and an 'error',
but the check would have to work for all possible sizes of 'Int' and 'CLong'
and would certainly require to convert from and to Integer at some point.
This is all error-prone and unsafe.
In contrast to that, 'show' is simple and safe.